% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_utils.R
\name{get_api_key}
\alias{get_api_key}
\title{Get Zotero API key}
\usage{
get_api_key(name)
}
\arguments{
\item{name}{A \code{character} string. This specifies which of your API keys to be used. Even for a public library, \code{name} is required.}
}
\value{
Zotero API key. Either a \code{character} (private library) string or \code{NULL} (public library).
}
\description{
Get Zotero API key specified with the \code{name}.
}
\details{
This function calls \code{load_api_key()} first. \code{load_api_key()} tries to load the cached API key specified with the \code{name} from a cache directory for \code{ztr4r} (see \code{api_key_path()} for its path). If \code{load_api_key()} successfully loads the API key, it will be used for the current session. Otherwise, \code{create_api_key()} will be called successively. It will create a new API key set by asking several user inputs, and store it with the \code{name} into the cache directory. The stored API key will be used for the current session.
}
\examples{
get_api_key(name = "your_api_key_name")
}
