% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_utils.R
\name{get_api_version}
\alias{get_api_version}
\title{Get Zotero API version}
\usage{
get_api_version()
}
\value{
Zotero API version in \code{integer}.
}
\description{
Get Zotero API version. Currently, the version 3 is supported.
}
\examples{
get_api_version()
}
