% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_utils.R
\name{load_api_key}
\alias{load_api_key}
\title{Load a cached Zotero Web API key}
\usage{
load_api_key(name)
}
\arguments{
\item{name}{A \code{character} string.}
}
\value{
Either a \code{list} that contains the API key or \code{NULL}.
}
\description{
Load a cached Zotero Web API key
}
\examples{
\dontrun{
load_api_key(name = "your_api_key_name")
}
}
