% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_request.R
\name{request}
\alias{request}
\title{Send request}
\usage{
request(name, verb, url, item, ..., timeout = 20)
}
\arguments{
\item{name}{A \code{character} string. Name of Zotero Web API key.}

\item{verb}{A \code{character} string. Either \code{"GET"} or \code{"POST"} depending on the request.}

\item{url}{A \code{character} string. URL to be set.}

\item{item}{A \code{logical}. Specify whether item request or not.}

\item{...}{Query inputs. For valid queries, please see \code{validate_query()}.}

\item{timeout}{An \code{integer}. Number of seconds to wait for a response until giving up.}
}
\value{
A \code{response()} object
}
\description{
Send request to Zotero Web API.
}
\examples{
\dontrun{
request(name = "your_api_key_name", verb = "GET", url = "url", item = TRUE)
}
}
