% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_utils.R
\name{save_api_key}
\alias{save_api_key}
\title{Save Zotero Web API key}
\usage{
save_api_key(name, auth)
}
\arguments{
\item{name}{A \code{character} string.}

\item{auth}{A \code{list} created by \code{create_api_key()}.}
}
\value{
\code{NULL}
}
\description{
Save Zotero Web API key
}
\examples{
\dontrun{
save_api_key(name = "your_api_key_name", auth = auth)
}
}
