% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_utils.R
\name{set_headers}
\alias{set_headers}
\title{Set HTTP headers.}
\usage{
set_headers(name)
}
\arguments{
\item{name}{A \code{character} string. This specifies which of your API keys to be used. Even for a public library, \code{name} is required. \code{name} is passed to \code{get_api_key()} function.}
}
\value{
\code{httr::add_headers()} with "Zotero-API-Version" and "Zotero-API-Key".
}
\description{
Set HTTP headers that used for sending requests.
}
\details{
Specifically, the returned HTTP header includes (1) Zotero API version and (2) Zotero API key. Please see \code{get_api_version()} and \code{get_api_key()} for details, respectively.
}
\examples{
set_headers(name = "your_api_key_name")
}
