% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Zotero.R
\name{Zotero}
\alias{Zotero}
\title{Zotero Class Object}
\description{
\code{Zotero} object stores
\itemize{
\item user information
\item cached requests
} and provides
\itemize{
\item \code{print} method
\item \code{request} method
}
}
\details{
The \code{request} method are wrapped with
\itemize{
\item \verb{ztr_read_*}
\item \verb{ztr_write_*}
\item \verb{ztr_search_*}
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{cache}}{Cached request table}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Zotero$new()}}
\item \href{#method-print}{\code{Zotero$print()}}
\item \href{#method-request}{\code{Zotero$request()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{Zotero} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Zotero$new(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Name of Zotero Web API key}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{Zotero} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Print a \code{Zotero} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Zotero$print()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
None
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-request"></a>}}
\if{latex}{\out{\hypertarget{method-request}{}}}
\subsection{Method \code{request()}}{
Send request
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Zotero$request(verb, path, ..., timeout = 20)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{verb}}{Request verb. Either \code{"GET"} or \code{"POST"}.}

\item{\code{path}}{Request path.}

\item{\code{...}}{Query inputs.}

\item{\code{timeout}}{Seconds for the timeout of request.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
This function is not supposed to be called directory. Instead, please use other specific functions. For available values for \code{path} and \code{...} (query), please see \url{https://www.zotero.org/support/dev/web_api/v3/basics}
}

}
}
