% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ztr4r-package.R
\docType{package}
\name{ztr4r-package}
\alias{ztr4r}
\alias{ztr4r-package}
\title{ztr4r: Zotero Web API wrapper and tool set for R}
\description{
ztr4r provides users Zotero Web API wrapper in R and a tool set for manipulating the contents in a Zotero library.
}
\examples{
\dontrun{
library(ztr4r)
}
}
\author{
\strong{Maintainer}: Teruaki Kido \email{tellurium3262@gmail.com}

}
\keyword{internal}
