% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_read.R
\name{ztr_read_collection}
\alias{ztr_read_collection}
\title{Read a specific collection or sub-collections within a specific collection}
\usage{
ztr_read_collection(zotero_object, collection, sub = FALSE, ..., timeout = 20)
}
\arguments{
\item{zotero_object}{\code{Zotero} object}

\item{collection}{Collection key}

\item{sub}{Specify whether to retrieve sub-collections within a specific collection. Otherwise, a specific collection will be retrieved.}

\item{...}{Query inputs. See \strong{Read Requests} and \strong{Sorting and Pagination} in the \href{https://www.zotero.org/support/dev/web_api/v3/basics}{Zotero Web API Documents} for available query.}

\item{timeout}{Seconds for the timeout of request (default: \code{20}).}
}
\value{
\code{httr::response} object
}
\description{
Read a specific collection or sub-collections within a specific collection
}
\examples{
\dontrun{
ztr <- Zotero$new("ztr4r")

# simple example
ztr \%>\%
  ztr_read_collection(collection = "foo")

# example with query
ztr \%>\%
  ztr_read_collection(
    collection = "foo", sub = TRUE,
    format = "json", sort = "dateAdded"
  )
}
}
