% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_read.R
\name{ztr_read_collection_tags}
\alias{ztr_read_collection_tags}
\title{Read Tags Within A Specific Collection In The Library}
\usage{
ztr_read_collection_tags(zotero_object, collection, ..., timeout = 20)
}
\arguments{
\item{zotero_object}{\code{Zotero} object.}

\item{collection}{Collection key.}

\item{...}{Query inputs. See \strong{Read Requests} and \strong{Sorting and Pagination} in the \href{https://www.zotero.org/support/dev/web_api/v3/basics}{Zotero Web API Documents} for available query.}

\item{timeout}{Seconds for the timeout of request (default: \code{20}).}
}
\value{
\code{httr::response} object
}
\description{
Read Tags Within A Specific Collection In The Library
}
\examples{
\dontrun{
ztr <- Zotero$new("ztr4r")

# simple example
ztr \%>\%
  ztr_read_collection_tags(collection = "foo")

# example with query
ztr \%>\%
  ztr_read_collection_tags(
    collection = "foo",
    format = "json", limit = 25
  )
}
}
