% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_read.R
\name{ztr_read_collections}
\alias{ztr_read_collections}
\title{Read (Top-Level) Collections In The Library}
\usage{
ztr_read_collections(zotero_object, top = FALSE, ..., timeout = 20)
}
\arguments{
\item{zotero_object}{\code{Zotero} object.}

\item{top}{Collection level (default: \code{FALSE}):
\itemize{
\item \code{FALSE}: All the collections in the library
\item \code{TRUE }: \strong{Top-level} collections in the library
}}

\item{...}{Query inputs. See \strong{Read Requests} and \strong{Sorting and Pagination} in the \href{https://www.zotero.org/support/dev/web_api/v3/basics}{Zotero Web API Documents} for available query.}

\item{timeout}{Seconds for the timeout of request (default: \code{20}).}
}
\value{
\code{httr::response} object
}
\description{
Read (Top-Level) Collections In The Library
}
\examples{
\dontrun{
ztr <- Zotero$new("ztr4r")

# simple example
ztr \%>\%
  ztr_read_collections()

# example with query
ztr \%>\%
  ztr_read_collections(
    top = TRUE,
    format = "json", sort = "dateModified", direction = "asc", limit = 25, start = 0
  )
}
}
