% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_read.R
\name{ztr_read_groups}
\alias{ztr_read_groups}
\title{Read Groups Assigned to A Specific User}
\usage{
ztr_read_groups(zotero_object, timeout = 20)
}
\arguments{
\item{zotero_object}{\code{Zotero} object}

\item{timeout}{Seconds for the timeout of request (default: \code{20})}
}
\value{
\code{httr::response} object
}
\description{
Read Groups Assigned to A Specific User
}
\examples{
\dontrun{
ztr <- Zotero$new("ztr4r")
ztr \%>\%
  ztr_read_groups()
}
}
