% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_read.R
\name{ztr_read_item}
\alias{ztr_read_item}
\title{Read a specific item or child items within a specific item}
\usage{
ztr_read_item(zotero_object, item, child = FALSE, ..., timeout = 20)
}
\arguments{
\item{zotero_object}{\code{Zotero} object}

\item{item}{Item key}

\item{child}{Specify whether to retrieve child items within a specific item Otherwise, a specific item will be retrieved.}

\item{...}{Query inputs. See \strong{Read Requests} and \strong{Sorting and Pagination} in the \href{https://www.zotero.org/support/dev/web_api/v3/basics}{Zotero Web API Documents} for available query.}

\item{timeout}{Seconds for the timeout of request (default: \code{20}).}
}
\value{
\code{httr::response} object
}
\description{
Read a specific item or child items within a specific item
}
\examples{
\dontrun{
ztr <- Zotero$new("ztr4r")

# simple example
ztr \%>\%
  ztr_read_item(item = "foo")

# example with query
ztr \%>\%
  ztr_read_item(
    item = "foo", child = TRUE,
    format = "json", sort = "dateAdded"
  )
}
}
