% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_read.R
\name{ztr_read_items}
\alias{ztr_read_items}
\title{Read items}
\usage{
ztr_read_items(
  zotero_object,
  collection = NULL,
  top = FALSE,
  ...,
  timeout = 20
)
}
\arguments{
\item{zotero_object}{\code{Zotero} object}

\item{collection}{Collection key. If specified, items within a specific collection will be retrieved.}

\item{top}{Specify item levels (default: \code{FALSE})
\itemize{
\item \code{FALSE}: All items in the library, excluding trashed items
\item \code{TRUE }: \strong{Top-level} items in the library, excluding trashed items.
}}

\item{...}{Query inputs. See \strong{Read Requests} and \strong{Sorting and Pagination} in the \href{https://www.zotero.org/support/dev/web_api/v3/basics}{Zotero Web API Documents} for available query.}

\item{timeout}{Seconds for the timeout of request (default: \code{20}).}
}
\value{
\code{httr::response} object
}
\description{
Read items
}
\examples{
\dontrun{
ztr <- Zotero$new("ztr4r")

# simple example
ztr \%>\%
  ztr_read_items()

# example with query: retrieve items
ztr \%>\%
  ztr_read_items(
    collection = NULL, top = TRUE,
    format = "json", sort = "dateAdded"
  )

# example with query: retrieve items in a specific collection
ztr \%>\%
  ztr_read_items(
    collection = "foo", top = TRUE,
    format = "json", sort = "dateAdded"
  )
}
}
