% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_read.R
\name{ztr_read_items_tags}
\alias{ztr_read_items_tags}
\title{Read Tags Assigned To Items (In A Given Collection)}
\usage{
ztr_read_items_tags(
  zotero_object,
  collection = NULL,
  top = FALSE,
  ...,
  timeout = 20
)
}
\arguments{
\item{zotero_object}{\code{Zotero} object.}

\item{collection}{Collection key (default: \code{NULL}).}

\item{top}{Item level (default: \code{FALSE}):
\itemize{
\item \code{FALSE}: All tags assigned to the items in the library/specific collection
\item \code{TRUE }: All tags assigned to the \strong{top-level} items in the library/specific collection
}}

\item{...}{Query inputs. See \strong{Read Requests} and \strong{Sorting and Pagination} in the \href{https://www.zotero.org/support/dev/web_api/v3/basics}{Zotero Web API Documents} for available query.}

\item{timeout}{Seconds for the timeout of request (default: \code{20}).}
}
\value{
\code{httr::response} object
}
\description{
Read
\enumerate{
\item All Tags In The Library, With The Ability To Filter Based On The Items
\item Tags Assigned To Top-Level Items
\item Tags Assigned To Items In A Given Collection
\item Tags Assigned To Top-Level Items In A Given Collection
}
}
\examples{
\dontrun{
ztr <- Zotero$new("ztr4r")

# simple example
ztr \%>\%
  ztr_read_items_tags()

# example with query: all tags in the library
ztr \%>\%
  ztr_read_items_tags(
    collection = NULL, top = FALSE,
    format = "json", limit = 25
  )

# example with query: tags assigned to top-level items
ztr \%>\%
  ztr_read_items_tags(
    collection = NULL, top = TRUE,
    format = "json", limit = 25
  )

# example with query: tags assigned to items in a given collection
ztr \%>\%
  ztr_read_items_tags(
    collection = "foo", top = FALSE,
    format = "json", limit = 25
  )


# example with query: tags assigned to top-level items in a given collection
ztr \%>\%
  ztr_read_items_tags(
    collection = "foo", top = TRUE,
    format = "json", limit = 25
  )
}
}
