% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_read.R
\name{ztr_read_search}
\alias{ztr_read_search}
\title{Read A Specific Saved Search In The Library}
\usage{
ztr_read_search(zotero_object, search, timeout = 20)
}
\arguments{
\item{zotero_object}{\code{Zotero} object.}

\item{search}{Search key.}

\item{timeout}{Seconds for the timeout of request (default: \code{20}).}
}
\value{
\code{httr::response} object
}
\description{
Read A Specific Saved Search In The Library
}
\examples{
\dontrun{
ztr <- Zotero$new("ztr4r")

# simple example
ztr \%>\%
  ztr_read_search()
}
}
