% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_read.R
\name{ztr_read_tags}
\alias{ztr_read_tags}
\title{Read All Tags In The Library Or Tags Of All Types Matching A Specific Name}
\usage{
ztr_read_tags(zotero_object, tag = NULL, ..., timeout = 20)
}
\arguments{
\item{zotero_object}{\code{Zotero} object.}

\item{tag}{URL encoded tag}

\item{...}{Query inputs. See \strong{Read Requests} and \strong{Sorting and Pagination} in the \href{https://www.zotero.org/support/dev/web_api/v3/basics}{Zotero Web API Documents} for available query.}

\item{timeout}{Seconds for the timeout of request (default: \code{20}).}
}
\value{
\code{httr::response} object
}
\description{
Read All Tags In The Library Or Tags Of All Types Matching A Specific Name
}
\examples{
\dontrun{
ztr <- Zotero$new("ztr4r")

# simple example
ztr \%>\%
  ztr_read_tags(tag = "tag=foo")

# example with query
ztr \%>\%
  ztr_read_tags(
    tag = "foo",
    format = "json", limit = 25
  )
}
}
