% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_utils.R
\name{api_key_path}
\alias{api_key_path}
\title{Get Zotero Web API key path}
\usage{
api_key_path(name)
}
\arguments{
\item{name}{A \code{character} string.}
}
\value{
A \code{character} string.
}
\description{
Get Zotero Web API key path
}
\examples{
\dontrun{
api_key_path(name = "your_api_key_name")
}
}
