% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_utils.R
\name{create_api_key}
\alias{create_api_key}
\title{Create a new Zotero Web API key}
\usage{
create_api_key(name)
}
\arguments{
\item{name}{A \code{character} string.}
}
\value{
A \code{list} that contains the API key.
}
\description{
Create a new Zotero Web API key
}
\examples{
\dontrun{
create_api_key(name = "your_api_key_name")
}
}
