% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_utils.R
\name{get_api_key}
\alias{get_api_key}
\title{Get Zotero Web API key}
\usage{
get_api_key(name)
}
\arguments{
\item{name}{A \code{character} string.}
}
\value{
Zotero Web API key in either a \code{character} string (private library) or \code{NULL} (public library).
}
\description{
Get Zotero Web API key
}
\examples{
\dontrun{
get_api_key(name = "your_api_key_name")
}
}
