% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_utils.R
\name{get_api_version}
\alias{get_api_version}
\title{Get Zotero Web API version}
\usage{
get_api_version()
}
\value{
Zotero Web API version in \code{integer}.
}
\description{
Get Zotero Web API version
}
\examples{
\dontrun{
get_api_version()
}
}
