% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_utils.R
\name{save_api_key}
\alias{save_api_key}
\title{Save API key}
\usage{
save_api_key(name, auth)
}
\arguments{
\item{name}{A \code{character} string. This specifies the name of a new API key. Even for a public library, \code{name} is required.}

\item{auth}{A \code{list} created in the \code{create_api_key()}.}
}
\value{
\code{NULL}
}
\description{
Save an API key in the user cache directory
}
\details{
This function saves the API key (\code{list}) created in the \code{create_api_key()} under the user cache directory in the \code{.rds} file format.
}
\examples{
save_api_key(name = "your_api_key_name", auth = auth)
}
