% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_utils.R
\name{set_headers}
\alias{set_headers}
\title{Set HTTP headers.}
\usage{
set_headers(name)
}
\arguments{
\item{name}{A \code{character} string.}
}
\value{
\code{httr::add_headers()}
}
\description{
Set HTTP headers.
}
\examples{
\dontrun{
set_headers(name = "your_api_key_name")
}
}
