% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_read.R
\name{ztr_read_items_tags}
\alias{ztr_read_items_tags}
\title{Read all tags assigned to items in the library or tags assigned to top-level items}
\usage{
ztr_read_items_tags(
  zotero_object,
  collection = NULL,
  top = FALSE,
  ...,
  timeout = 20
)
}
\arguments{
\item{zotero_object}{\code{Zotero} object}

\item{collection}{Collection key. Tags within a specific collection in the library will be retrieved.}

\item{top}{Specify item levels (default: \code{FALSE})
\itemize{
\item \code{FALSE}: All tags in the library, with the ability to filter based on the items
\item \code{TRUE }: Tags assigned to top-level items
}}

\item{...}{Query inputs. See \strong{Read Requests} and \strong{Sorting and Pagination} in the \href{https://www.zotero.org/support/dev/web_api/v3/basics}{Zotero Web API Documents} for available query.}

\item{timeout}{Seconds for the timeout of request (default: \code{20}).}
}
\value{
\code{httr::response} object
}
\description{
Read all tags assigned to items in the library or tags assigned to top-level items
}
\examples{
\dontrun{
ztr <- Zotero$new("ztr4r")

# simple example
ztr \%>\%
  ztr_read_items_tags()

# example with query: retrieve items
ztr \%>\%
  ztr_read_items_tags(
    collection = NULL, top = TRUE,
    format = "json", sort = "dateAdded"
  )

# example with query: retrieve items in a specific collection
ztr \%>\%
  ztr_read_items_tags(
    collection = "foo", top = TRUE,
    format = "json", sort = "dateAdded"
  )
}
}
